+++ src/main/java/org/danji/auth/account/dto/AuthResultDTO.java
@@ -1,0 +1,13 @@
+package org.danji.auth.account.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthResultDTO {
+    String token;
+    UserInfoDTO user;
+}
