+++ src/main/java/org/danji/auth/account/dto/LoginDTO.java
@@ -1,0 +1,28 @@
+package org.danji.auth.account.dto;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.security.authentication.BadCredentialsException;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class LoginDTO {
+    private String username;
+    private String password;
+
+
+    public static LoginDTO of(HttpServletRequest request) {
+        ObjectMapper om = new ObjectMapper();
+        try {
+            return om.readValue(request.getInputStream(), LoginDTO.class);
+        } catch (IOException e) {
+            throw new BadCredentialsException("username 또는 password가 없음");
+        } //catch
+    } //of
+} //class
