+++ src/main/java/org/danji/auth/handler/LoginSuccessHandler.java
@@ -1,0 +1,49 @@
+package org.danji.auth.handler;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import org.danji.auth.account.domain.CustomUser;
+import org.danji.auth.account.dto.AuthResultDTO;
+import org.danji.auth.account.dto.UserInfoDTO;
+import org.danji.global.security.jwt.util.JsonResponse;
+import org.danji.global.security.jwt.util.JwtProcessor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Log4j2
+@Component
+@RequiredArgsConstructor
+public class LoginSuccessHandler implements AuthenticationSuccessHandler {
+    private final JwtProcessor jwtProcessor;
+
+    private AuthResultDTO makeAuthResult(CustomUser user) {
+
+        //JsonResponse로 보낼 값들 만ㄷ르어야함.
+        //성공했으므로  Authentication객체가 이미 만들어져서
+        //SecuriyContextHolder에 들어가 있음.
+
+        String username = user.getUsername();
+        // 토큰 생성
+        String token = jwtProcessor.generateToken(username);
+        // 토큰 + 사용자 기본 정보 (사용자명, ...)를 묶어서 AuthResultDTO 구성
+        return new AuthResultDTO(token, UserInfoDTO.of(user.getMember()));
+    }
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+                                        Authentication authentication) throws IOException, ServletException {
+        // 인증 결과 Principal
+        CustomUser user = (CustomUser) authentication.getPrincipal();
+
+        // 인증 성공 결과를 JSON으로 직접 응답
+        AuthResultDTO result = makeAuthResult(user);
+        JsonResponse.send(response, result);
+    }
+}
+
