--- src/main/java/org/danji/config/RootConfig.java
@@ -1,76 +1,0 @@
-package org.danji.config;
-
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
-import lombok.extern.log4j.Log4j2;
-import org.apache.ibatis.session.SqlSessionFactory;
-import org.mybatis.spring.SqlSessionFactoryBean;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import javax.sql.DataSource;
-
-@Configuration
-@PropertySource({"classpath:/application.properties"})
-@MapperScan(basePackages = {"org.danji.board.mapper",
-        "org.danji.member.mapper",
-        "org.danji.travel.mapper"
-})
-@ComponentScan(basePackages = {"org.danji.board.service",
-        "org.danji.member.service",
-        "org.danji.travel.service"
-})
-@Log4j2
-@EnableTransactionManagement
-public class RootConfig {
-
-    @Value("${jdbc.driver}")
-    String driver;
-    @Value("${jdbc.url}")
-    String url;
-    @Value("${jdbc.username}")
-    String username;
-    @Value("${jdbc.password}")
-    String password;
-    @Autowired
-    ApplicationContext applicationContext;
-
-    @Bean
-    public DataSource dataSource() {
-        HikariConfig config = new HikariConfig();
-
-        config.setDriverClassName(driver);
-        config.setJdbcUrl(url);
-        config.setUsername(username);
-        config.setPassword(password);
-
-        HikariDataSource dataSource = new HikariDataSource(config);
-        return dataSource;
-    }
-
-    @Bean
-    public SqlSessionFactory sqlSessionFactory() throws Exception {
-        SqlSessionFactoryBean sqlSessionFactory = new SqlSessionFactoryBean();
-        sqlSessionFactory.setConfigLocation(
-                applicationContext.getResource("classpath:/mybatis-config.xml"));
-        sqlSessionFactory.setDataSource(dataSource());
-
-        return sqlSessionFactory.getObject();
-    }
-
-    @Bean
-    public DataSourceTransactionManager transactionManager() {
-        DataSourceTransactionManager manager = new DataSourceTransactionManager(dataSource());
-
-        return manager;
-    }
-
-}
