--- src/main/java/org/danji/config/ServletConfig.java
@@ -1,64 +1,0 @@
-package org.danji.config;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.web.multipart.MultipartResolver;
-import org.springframework.web.multipart.support.StandardServletMultipartResolver;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@EnableWebMvc
-@ComponentScan(basePackages = {
-        "org.danji.exception",
-        "org.danji.controller",
-        "org.danji.board.controller",
-        "org.danji.member.controller",
-        "org.danji.travel.controller",
-})
-public class ServletConfig implements WebMvcConfigurer {
-
-    // 선언적 코드(java)
-    @Override
-    public void addViewControllers(ViewControllerRegistry registry) {
-        registry.addViewController("/")
-                .setViewName("redirect:/swagger-ui.html");
-    }
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        // Swagger UI 리소스를 위한 핸들러 설정
-        registry.addResourceHandler("/swagger-ui.html")
-                .addResourceLocations("classpath:/META-INF/resources/");
-
-        // Swagger WebJar 리소스 설정
-        registry.addResourceHandler("/webjars/**")
-                .addResourceLocations("classpath:/META-INF/resources/webjars/");
-
-        // Swagger 리소스 설정
-        registry.addResourceHandler("/swagger-resources/**")
-                .addResourceLocations("classpath:/META-INF/resources/");
-
-        registry.addResourceHandler("/v2/api-docs")
-                .addResourceLocations("classpath:/META-INF/resources/");
-
-        registry
-                .addResourceHandler("/resources/**")     // url이 /resources/로 시작하는 모든 경로
-                .addResourceLocations("/resources/");    // webapp/resources/경로로 매핑
-
-        registry.addResourceHandler("/assets/**")
-                .addResourceLocations("/resources/assets/");
-    }
-
-
-    //	Servlet 3.0 파일 업로드 사용시
-    @Bean
-    public MultipartResolver multipartResolver() {
-        StandardServletMultipartResolver resolver = new StandardServletMultipartResolver();
-        return resolver;
-    }
-
-
-}
