--- src/main/java/org/danji/config/SwaggerConfig.java
@@ -1,72 +1,0 @@
-package org.danji.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.bind.annotation.RestController;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.ApiKey;
-import springfox.documentation.service.AuthorizationScope;
-import springfox.documentation.service.SecurityReference;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spi.service.contexts.SecurityContext;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.List;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-
-    private final String API_NAME = "Board API";
-    private final String API_VERSION = "1.0";
-    private final String API_DESCRIPTION = "Board API 명세서";
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
-                .title(API_NAME)
-                .description(API_DESCRIPTION)
-                .version(API_VERSION)
-                .build();
-
-    }
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                // JWT 인증
-                .securityContexts(List.of(this.securityContext())) // SecurityContext 설정
-                .securitySchemes(List.of(this.apiKey())) // ApiKey 설정
-                .select()
-                .apis(RequestHandlerSelectors.withClassAnnotation(RestController.class))
-                .paths(PathSelectors.any())
-                .build()
-                .apiInfo(apiInfo());
-
-    }
-
-    // JWT SecurityContext 구성
-    private SecurityContext securityContext() {
-        return SecurityContext.builder()
-
-                .securityReferences(defaultAuth())
-                .build();
-
-    }
-
-    private List<SecurityReference> defaultAuth() {
-        AuthorizationScope authorizationScope = new AuthorizationScope("global",
-                "accessEverything");
-        AuthorizationScope[] authorizationScopes = new AuthorizationScope[1];
-        authorizationScopes[0] = authorizationScope;
-        return List.of(new SecurityReference("Authorization", authorizationScopes));
-    }
-
-    // ApiKey 정의
-    private ApiKey apiKey() {
-        return new ApiKey("Authorization", "Authorization", "header");
-    }
-}
