--- src/main/java/org/danji/exception/CommonExceptionAdvice.java
@@ -1,20 +1,0 @@
-package org.danji.exception;
-
-
-import lombok.extern.log4j.Log4j2;
-import org.springframework.core.annotation.Order;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.servlet.NoHandlerFoundException;
-
-@ControllerAdvice
-@Log4j2
-@Order(1)
-public class CommonExceptionAdvice {
-
-    @ExceptionHandler(NoHandlerFoundException.class)
-    public String handle404(NoHandlerFoundException ex) {
-        return "/resources/index.html";
-    }
-
-}
