+++ src/main/java/org/danji/global/config/SecurityInitializer.java
@@ -1,0 +1,23 @@
+package org.danji.global.config;
+
+import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
+import org.springframework.web.filter.CharacterEncodingFilter;
+import org.springframework.web.multipart.support.MultipartFilter;
+
+import javax.servlet.ServletContext;
+
+public class SecurityInitializer extends AbstractSecurityWebApplicationInitializer {
+    // 문자셋 필터
+    private CharacterEncodingFilter encodingFilter() {
+        CharacterEncodingFilter encodingFilter = new CharacterEncodingFilter();
+        encodingFilter.setEncoding("UTF-8");
+        encodingFilter.setForceEncoding(true);
+        return encodingFilter;
+    }
+
+    @Override
+    protected void beforeSpringSecurityFilterChain(ServletContext servletContext) {
+        insertFilters(servletContext, encodingFilter(), new MultipartFilter());
+    }
+
+}
