+++ src/main/java/org/danji/global/security/jwt/filter/AuthenticationErrorFilter.java
@@ -1,0 +1,34 @@
+package org.danji.global.security.jwt.filter;
+
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.MalformedJwtException;
+import io.jsonwebtoken.UnsupportedJwtException;
+import io.jsonwebtoken.security.SignatureException;
+import org.danji.global.security.jwt.util.JsonResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+
+@Component
+public class AuthenticationErrorFilter extends OncePerRequestFilter {
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+            throws ServletException, IOException {
+        try {
+            super.doFilter(request, response, filterChain);
+        } catch (ExpiredJwtException e) {
+            JsonResponse.sendError(response, HttpStatus.UNAUTHORIZED, "토큰의 유효시간이 지났습니다.");
+        } catch (UnsupportedJwtException | MalformedJwtException | SignatureException e) {
+            JsonResponse.sendError(response, HttpStatus.UNAUTHORIZED, e.getMessage());
+        } catch (ServletException e) {
+            JsonResponse.sendError(response, HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+        }
+    }
+}
