+++ src/main/java/org/danji/global/security/jwt/util/JsonResponse.java
@@ -1,0 +1,26 @@
+package org.danji.global.security.jwt.util;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.http.HttpStatus;
+
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.Writer;
+
+public class JsonResponse {
+    public static <T> void send(HttpServletResponse response, T result) throws IOException {
+        ObjectMapper om = new ObjectMapper();
+        response.setContentType("application/json;charset=UTF-8");
+        Writer out = response.getWriter();
+        out.write(om.writeValueAsString(result));
+        out.flush();
+    }
+
+    public static void sendError(HttpServletResponse response, HttpStatus status, String message) throws IOException {
+        response.setStatus(status.value());
+        response.setContentType("application/json;charset=UTF-8");
+        Writer out = response.getWriter();
+        out.write(message);
+        out.flush();
+    }
+}
