+++ src/main/java/org/danji/global/security/jwt/util/JwtProcessor.java
@@ -1,0 +1,52 @@
+package org.danji.global.security.jwt.util;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.stereotype.Component;
+
+import java.nio.charset.StandardCharsets;
+import java.security.Key;
+import java.util.Date;
+
+@Component
+public class JwtProcessor {
+
+    //1. jwt 생성
+    static private final long TOKEN_VALID_MILISECOND = 1000L * 60 * 2; // 2분
+    private final String secretKey
+            = "충분히긴임의의(랜덤한) 비밀키문자열배정";
+    private final Key key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
+
+    //3. jwt유효성 검증(브라우저로부터 서버로 전송된 경우, 서버에서 검증할 때)
+    public boolean validateToken(String token) {
+        Jws<Claims> claims = Jwts.parserBuilder()
+                .setSigningKey(key)
+                .build()
+                .parseClaimsJws(token);
+        return true;
+    }
+
+    //private Key key = Keys.secretKeyFor(SignatureAlgorithm.HS256);  // 운영시 사용
+    // JWT 생성
+    public String generateToken(String subject) {
+        return Jwts.builder()
+                .setSubject(subject)
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(new Date().getTime() + TOKEN_VALID_MILISECOND))
+                .signWith(key)
+                .compact();
+    }
+
+    //2. username(id)추출
+    public String getUsername(String token) {
+        return Jwts.parserBuilder()
+                .setSigningKey(key)
+                .build()
+                .parseClaimsJws(token)
+                .getBody()
+                .getSubject();
+    }
+
+}
