+++ src/main/java/org/danji/member/dto/MemberDTO.java
@@ -4,8 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.danji.security.account.domain.AuthVO;
-import org.danji.security.account.domain.MemberVO;
+import org.danji.auth.account.domain.AuthVO;
+import org.danji.auth.account.domain.MemberVO;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Date;
