+++ src/main/java/org/danji/member/dto/MemberJoinDTO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.danji.security.account.domain.MemberVO;
+import org.danji.auth.account.domain.MemberVO;
 import org.springframework.web.multipart.MultipartFile;
 
 @NoArgsConstructor
