+++ src/main/java/org/danji/member/mapper/MemberMapper.java
@@ -1,8 +1,9 @@
 package org.danji.member.mapper;
 
+import org.danji.auth.account.domain.AuthVO;
+import org.danji.auth.account.domain.MemberVO;
 import org.danji.member.dto.ChangePasswordDTO;
-import org.danji.security.account.domain.AuthVO;
-import org.danji.security.account.domain.MemberVO;
+
 
 public interface MemberMapper {
