--- src/main/java/org/danji/security/account/dto/AuthResultDTO.java
@@ -1,13 +1,0 @@
-package org.danji.security.account.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthResultDTO {
-    String token;
-    UserInfoDTO user;
-}
