--- src/main/java/org/danji/security/account/dto/UserInfoDTO.java
@@ -1,27 +1,0 @@
-package org.danji.security.account.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.danji.security.account.domain.MemberVO;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserInfoDTO {
-    String username;
-    String email;
-    List<String> roles;
-
-    public static UserInfoDTO of(MemberVO member) {
-        return new UserInfoDTO(
-                member.getUsername(),
-                member.getEmail(),
-                member.getAuthList().stream()
-                        .map(a -> a.getAuth())
-                        .toList()
-        );
-    }
-}
