--- src/main/java/org/danji/security/config/SecurityConfig.java
@@ -1,148 +1,0 @@
-package org.danji.security.config;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.danji.security.filter.AuthenticationErrorFilter;
-import org.danji.security.filter.JwtAuthenticationFilter;
-import org.danji.security.filter.JwtUsernamePasswordAuthenticationFilter;
-import org.danji.security.handler.CustomAccessDeniedHandler;
-import org.danji.security.handler.CustomAuthenticationEntryPoint;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
-
-@Configuration
-@EnableWebSecurity
-@Log4j2
-@MapperScan(basePackages = {"org.danji.security.account.mapper"})
-@ComponentScan(basePackages = {"org.danji.security"})
-@RequiredArgsConstructor
-
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    private final UserDetailsService userDetailsService;
-    private final JwtAuthenticationFilter jwtAuthenticationFilter;
-    private final AuthenticationErrorFilter authenticationErrorFilter;
-    private final CustomAccessDeniedHandler accessDeniedHandler;
-    private final CustomAuthenticationEntryPoint authenticationEntryPoint;
-
-    @Autowired
-    private JwtUsernamePasswordAuthenticationFilter jwtUsernamePasswordAuthenticationFilter;
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    // 문자셋필터
-    // post방식의 전달시 body에 들어있는 값 한글 인코딩 필터
-//    public CharacterEncodingFilter encodingFilter() {
-//        CharacterEncodingFilter encodingFilter = new CharacterEncodingFilter();
-//        encodingFilter.setEncoding("UTF-8");
-//        encodingFilter.setForceEncoding(true);
-//        return encodingFilter;
-//    }
-
-    // AuthenticationManager 빈 등록
-    @Bean
-//    JWT 방식은 폼로그인과달리Spring Security의기 본인증필터를사용하지않고,
-//    클라이언트→ JWT 토큰→ 커스텀필터
-//    (OncePerRequestFilter 등) → SecurityContext 직접 설정
-    public AuthenticationManager authenticationManager() throws Exception {
-        return super.authenticationManager();
-
-    }
-
-    // cross origin 접근 허용
-    @Bean
-    public CorsFilter corsFilter() {
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        CorsConfiguration config = new CorsConfiguration();
-        config.setAllowCredentials(true);
-        config.addAllowedOriginPattern("*");
-        config.addAllowedHeader("*");
-        config.addAllowedMethod("*");
-        source.registerCorsConfiguration("/**", config);
-        return new CorsFilter(source);
-    }
-
-    // 접근 제한무시경로설정–resource
-    @Override
-    public void configure(WebSecurity web) throws Exception {
-        web.ignoring().antMatchers(
-                "/assets/**",
-                "/*",
-//            "/api/member/**",
-                // Swagger 관련 url은 보안에서 제외
-                "/swagger-ui.html", "/webjars/**", "/swagger-resources/**", "/v2/api-docs");
-    }
-
-    @Override
-    public void configure(HttpSecurity http) throws Exception {
-        http
-//                .addFilterBefore(encodingFilter(), CsrfFilter.class)
-                // Jwt
-                //
-                .addFilterBefore(authenticationErrorFilter, UsernamePasswordAuthenticationFilter.class)
-                .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
-                .addFilterBefore(jwtUsernamePasswordAuthenticationFilter,
-                        UsernamePasswordAuthenticationFilter.class);
-
-        http.httpBasic().disable() // 기본 HTTP 인증비활성화
-                .csrf().disable() // CSRF 비활성화
-                .formLogin().disable()  // formLogin 비활성화- 관련 필터 해제
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS); // 세션 생성 모드 설정
-
-        http
-                .exceptionHandling()
-                .authenticationEntryPoint(authenticationEntryPoint)
-                .accessDeniedHandler(accessDeniedHandler);
-
-        http
-                .authorizeRequests() // 경로별 접근 권한 설정
-                .antMatchers(HttpMethod.OPTIONS).permitAll()
-                // 🌐 회원 관련 공개 API (인증 불필요)
-                .antMatchers(HttpMethod.GET, "/api/member/checkusername/**").permitAll()     // ID 중복 체크
-                .antMatchers(HttpMethod.POST, "/api/member").permitAll()                    // 회원가입
-                .antMatchers(HttpMethod.GET, "/api/member/*/avatar").permitAll()            // 아바타 이미지
-
-                // 회원 관련 인증 요구 경로
-                .antMatchers(HttpMethod.POST, "/api/member").authenticated() // 회원 등록
-                .antMatchers(HttpMethod.PUT, "/api/member", "/api/member/*/changepassword").authenticated() // 회원 정보 수정, 비밀번호 변경
-
-                // 게시판 관련 인증 요구 경로
-                .antMatchers(HttpMethod.POST, "/api/board/**").authenticated() // 쓰기
-                .antMatchers(HttpMethod.PUT, "/api/board/**").authenticated()  // 수정
-                .antMatchers(HttpMethod.DELETE, "/api/board/**").authenticated() // 삭제
-                .anyRequest().permitAll(); // 나머지 허용
-
-    }
-
-    @Override
-    protected void configure(AuthenticationManagerBuilder auth)
-            throws Exception {
-        log.info("configure .........................................");
-
-        auth.userDetailsService(userDetailsService)
-                .passwordEncoder(passwordEncoder());
-
-    }
-}
