--- src/main/java/org/danji/security/handler/CustomAuthenticationEntryPoint.java
@@ -1,24 +1,0 @@
-package org.danji.security.handler;
-
-import lombok.extern.log4j.Log4j2;
-import org.danji.security.util.JsonResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Log4j2
-@Component
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException)
-            throws IOException, ServletException {
-        log.error("========== ============");
-        JsonResponse.sendError(response, HttpStatus.UNAUTHORIZED, authException.getMessage());
-    }
-}
