--- src/main/java/org/danji/security/service/CustomUserDetailsService.java
@@ -1,25 +1,0 @@
-package org.danji.security.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.danji.security.account.domain.CustomUser;
-import org.danji.security.account.domain.MemberVO;
-import org.danji.security.account.mapper.UserDetailsMapper;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-@Component
-@Log4j2
-@RequiredArgsConstructor //생성자를 자동으로 주입해줌.
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private final UserDetailsMapper userDetailsMapper;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        MemberVO memberVO = userDetailsMapper.get(username);
-        return new CustomUser(memberVO);
-    }
-}
