+++ src/main/resources/org/danji/security/account/mapper/UserDetailsMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="org.danji.security.account.mapper.UserDetailsMapper">
+<mapper namespace="org.danji.auth.account.mapper.UserDetailsMapper">
     <resultMap id="authMap" type="AuthVO">
         <result property="username" column="username"/>
         <result property="auth" column="auth"/>
